requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1
django-import-export==2.8.0
django-mathfilters==1.0.0
docopt==0.6.2
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
lxml==4.9.1
MarkupPy==1.14
num2words==0.5.12
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.11
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
